+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Hide local timezone with `send-as-utc = true` in `accounts.conf`.
 - Persistent command history in `~/.cache/aerc/history`.
 - Cursor shape support in embedded terminals.
+- Bracketed paste support.
 
 ### Changed